Patch by Christian Ohm: 044 - Use int for gamma to avoid float imprecision
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@866 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/display.h
@@ -181,7 +181,7 @@ BOOL CheckInScrollLimits(SDWORD *xPos,SDWORD *zPos);
 extern BOOL CheckScrollLimits(void);
 //extern BOOL	widgetsOn;
 extern BOOL	rotActive;
-extern float	gammaValue;
+extern int gammaValue;
 //extern BOOL	forceWidgetsOn;
 
 BASE_OBJECT	*mouseTarget( void );
